Split verbatim into verbatim_cli and verbatim_serve
