+++ verbatim_cli/__init__.py
@@ -1,0 +1,5 @@
+"""CLI entrypoint for verbatim."""
+
+from verbatim_cli.main import main
+
+__all__ = ["main"]
