+++ verbatim_serve/__init__.py
@@ -1,0 +1,5 @@
+"""HTTP server entrypoint for verbatim."""
+
+from verbatim_serve.main import main
+
+__all__ = ["main"]
